Implementing small functions and changes in preparation for new grocery shopping analytics features.
The plan: The user will be able to initiate a new activity for their currently selected list and begin to shop for items inside their grocery list. The program will cycle through each object in a grocery list , prompting the user to enter actual cost and quantity purchased. The user will have the option to skip/cancel the item and move on to the next item. If skipped, the item will be added to back of the queue. If canceled, the item will be removed from the queue and have actual cost and quantity logged as zero.

 Once complete, the user will be able to view their shopping history with the specified list.

Modifed GroceryListListAdapter and appropriate xml files to display total cost of all items in main activity items.

GroceryItem class
-Added setQuantity(int) function GroceryItem class

GroceryList class
-Added removeGrocery(int):void function to facilitate the deletion of grocery objects from a list
-Added getTotalCost():double function to retrieve the total cost of all goods in a grocery list.

GroceryListViewActivity
-Added options menu to support future planned features for grocery shopping analytics
-XML: Added Buttons to support future planned features for grocery shopping analytics